gpa calculator .cpp
The user’s input is used to compute the CGPA, which
includes information like the number of courses taken and
the grades earned in each one. The program also shows each
student’s overall course grade. CGPA Calculator calculates a
student’s Cumulative Grade Point Average (CGPA) from the
given exam results. The program can show the individual
grades of each course, calculate total credits and total grade
points achieved, determine the GPA for the semester, and
based on all the data, it can generate and present the CGPA
of the student.